# api-docs.yml
openapi: 3.0.0
info:
  title: Product Service API
  version: 1.0.0
paths:
  /products:
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                userId:
                  type: string
              required:
                - id
                - name
                - userId
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  userId:
                    type: string
        '400':
          description: Bad Request
  /products/{id}:
    get:
      summary: Retrieve product details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  userId:
                    type: string
        '404':
          description: Product not found
  /products/user/{userId}:
    get:
      summary: Retrieve all products created by a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    userId:
                      type: string